회원가입 페이지 methods hooks 전환
